OnlineMetrics

Documentation for OnlineMetrics.

OnlineMetrics.AbstractMetric

OnlineMetrics.Accuracy

OnlineMetrics.AverageMeasure

OnlineMetrics.BinaryPrecision

OnlineMetrics.BinaryRecall

OnlineMetrics.ClassificationMetric

OnlineMetrics.ConfusionMatrix

OnlineMetrics.MIoU

OnlineMetrics.MetricCollection

OnlineMetrics.Precision

OnlineMetrics.Recall

OnlineMetrics.params

OnlineMetrics.step!

OnlineMetrics.value

OnlineMetrics.AbstractMetric — Type

Metrics are measures of a model's performance, such as loss, accuracy, or squared error.

Each metric must implement the following interface:

step!(metric, ŷ, y): Updates the metric state for the given batch of predictions and labels.

value(metric): Return the current value of the metric.

params(metric): Return a NamedTuple containing any internal parameters you wish to report.

Example Implementation

mutable struct Accuracy <: ClassificationMetric
    correct::Int
    n::Int
    Accuracy() = new(0,0)
end

function step!(m::Accuracy, ŷ::AbstractVector{<:Integer}, y::AbstractVector{<:Integer})
    m.correct += sum(ŷ .== y)
    m.n += length(ŷ)
end

value(m::Accuracy) = m.correct / max(m.n, 1)

params(x::Accuracy) = (;x.n, x.correct)

OnlineMetrics.Accuracy — Type

Accuracy()

Measures the model's overall accuracy as correct / n.

OnlineMetrics.AverageMeasure — Type

AverageMeasure(measure::Function)

Tracks the average of the given measure over mini-batches.

OnlineMetrics.BinaryPrecision — Type

BinaryPrecision()

A variant of Precision specialized for binary classification.

OnlineMetrics.BinaryRecall — Type

BinaryRecall()

A variant of Recall specialized for binary classification.

OnlineMetrics.ClassificationMetric — Type

Classification metrics are used to evaluate the performance of models that predict a discrete label for each observation. Subtypes should implement an batchstate method which assumes that both ŷ and y are encoded as integers.

OnlineMetrics.ConfusionMatrix — Type

ConfusionMatrix(nclasses::Int)

Calculate the confusion matrix over two or more classes. The columns of the resulting nclasses x nclasses matrix correspond to the true label while the rows correspond to the prediction.

Arguments

nclasses::Int: The number of possible classes in the classification task.

OnlineMetrics.MIoU — Type

MIoU(nclasses::Int)

Mean Intersection over Union (MIoU) is a measure of the overlap between a prediction and a label. This measure is frequently used for segmentation models.

OnlineMetrics.MetricCollection — Type

MetricCollection(metrics...)

An object to track one or more metrics concurrently.

Example

julia> mc = MetricCollection(Accuracy(), MIoU(2));

julia> step!(mc, [0, 0, 1, 0], [0, 0, 1, 1]);

julia> mc
MetricCollection
├─ Accuracy
│  ├─ :value ⇒ 0.75
│  ├─ :n ⇒ 4
│  └─ :correct ⇒ 3
└─ MIoU
   ├─ :value ⇒ 0.583333
   ├─ :union ⇒ [3, 2]
   └─ :intersection ⇒ [2, 1]

OnlineMetrics.Precision — Type

Precision(nclasses::Int; agg=:macro)

Precision is the ratio of true positives to the sum of true positives and false positives, measuring the accuracy of positive predictions.

Arguments

nclasses::Int: The number of classes for the classification task.

Keyword Arguments

agg: Specifies the type of precision aggregation to be computed. The possible values are:

:macro: Calculates macro-averaged precision, which computes the precision for each class independently and then takes the average.

:micro: Calculates micro-averaged precision, which aggregates the contributions of all classes to compute a single precision value.

:nothing: Calculates the per-class precision, which is returned as a Vector with the same length as classes.

OnlineMetrics.Recall — Type

Recall(nclasses::Int; agg=:macro)

Recall, also known as sensitivity or true positive rate, is the ratio of true positives to the sum of true positives and false negatives, measuring the ability of the classifier to identify all positive instances.

Arguments

nclasses::Int: The number of classes for the classification task.

Keyword Arguments

agg: Specifies the type of recall aggregation to be computed. The possible values are:

:macro: Calculates macro-averaged recall, which computes the recall for each class independently and then takes the average.

:micro: Calculates micro-averaged recall, which aggregates the contributions of all classes to compute a single recall value.

:nothing: Calculates the per-class recall, which is returned as a Vector with the same length as classes.

OnlineMetrics.params — Method

params(m::AbstractMetric)

Returns any internal parameters used to track the metric's current value.

OnlineMetrics.step! — Function

step!(m::AbstractMetric, ŷ, y)

Update the metric state for the given batch of labels and predictions.

OnlineMetrics.value — Function

value(m::AbstractMetric)

Compute the performance measure from the current metric state.